update to 7.1 and refactor
