+++ .env.example
@@ -1,0 +1,33 @@
+PROJECT_ID="eoapi"
+STAGE="production"
+TAGS='{"owner": "user_a"}'
+
+# Ingest Options
+AUTH_PROVIDER_JWKS_URL="https://auth.io"
+DATA_ACCESS_ROLE_ARN="arn:..."
+
+# Database Options
+DB_INSTANCE_TYPE="t3.micro"
+DB_ALLOCATED_STORAGE="5"
+PUBLIC_DB_SUBNET="FALSE"
+NAT_GATEWAY_COUNT="4"
+
+# Bastion Host Options
+BASTION_HOST="TRUE"
+BASTION_HOST_CREATE_ELASTIC_IP="TRUE"
+BASTION_HOST_ALLOW_IP_LIST='["http://0.0.0.0"]'
+BASTION_HOST_USER_DATA="path/file.txt"
+
+# Raster Options
+RASTER_BUCKETS='["raster-east"]'
+
+# STAC Browser Options
+STAC_BROWSER_VERSION="0.1.0"
+
+# API Gateway Options
+ACM_CERTIFICATE_ARN="arn:..."
+STAC_API_CUSTOM_DOMAIN="https://stac.api"
+RASTER_API_CUSTOM_DOMAIN="https://raster.api"
+VECTOR_API_CUSTOM_DOMAIN="https://vector.api"
+STAC_INGESTOR_API_CUSTOM_DOMAIN="https://ingestor.api"
+
