+++ docker-compose.yml
@@ -56,7 +56,7 @@ services:
     # At the time of writing, rasterio and psycopg wheels are not available for arm64 arch
     # so we force the image to be built with linux/amd64
     platform: linux/amd64
-    image: ghcr.io/stac-utils/titiler-pgstac:0.5.1
+    image: ghcr.io/stac-utils/titiler-pgstac:1.2.2
     ports:
       - "${MY_DOCKER_IP:-127.0.0.1}:8082:8082"
     environment:
@@ -102,7 +102,7 @@ services:
       - ./dockerfiles/scripts:/tmp/scripts
 
   vector:
-    image: ghcr.io/developmentseed/tipg:0.3.1
+    image: ghcr.io/developmentseed/tipg:0.6.3
     ports:
       - "${MY_DOCKER_IP:-127.0.0.1}:8083:8083"
     environment:
@@ -131,7 +131,7 @@ services:
       - ./dockerfiles/scripts:/tmp/scripts
 
   database:
-    image: ghcr.io/stac-utils/pgstac:v0.7.10
+    image: ghcr.io/stac-utils/pgstac:v0.8.5
     environment:
       - POSTGRES_USER=username
       - POSTGRES_PASSWORD=password
