+++ infrastructure/config.py
@@ -1,11 +1,10 @@
 from typing import Any, Dict, List, Optional, Union
 
-import pydantic
 import yaml
 from aws_cdk import aws_ec2
-from pydantic import Field
-from pydantic_core.core_schema import FieldValidationInfo
-from pydantic_settings import BaseSettings
+from pydantic import Field, ValidationInfo, field_validator, model_validator
+from pydantic_settings import BaseSettings, SettingsConfigDict
+from typing_extensions import Self
 
 
 class AppConfig(BaseSettings):
@@ -71,9 +70,9 @@ class AppConfig(BaseSettings):
         Ignored if `bastion_host` equals `False`.""",
         default=aws_ec2.UserData.for_linux(),
     )
-    titiler_buckets: List[str] = Field(
+    raster_buckets: List[str] = Field(
         description="""Path to YAML file containing list of
-        buckets to grant access to the titiler API""",
+        buckets to grant access to the Raster API""",
         default=[],
     )
     acm_certificate_arn: Optional[str] = Field(
@@ -87,18 +86,18 @@ class AppConfig(BaseSettings):
         Must provide `acm_certificate_arn`""",
         default=None,
     )
-    titiler_pgstac_api_custom_domain: Optional[str] = Field(
-        description="""Custom domain name for the titiler pgstac API.
+    raster_api_custom_domain: Optional[str] = Field(
+        description="""Custom domain name for the Raster API.
+        Must provide `acm_certificate_arn`""",
+        default=None,
+    )
+    vector_api_custom_domain: Optional[str] = Field(
+        description="""Custom domain name for the Vector API.
         Must provide `acm_certificate_arn`""",
         default=None,
     )
     stac_ingestor_api_custom_domain: Optional[str] = Field(
         description="""Custom domain name for the STAC ingestor API.
-        Must provide `acm_certificate_arn`""",
-        default=None,
-    )
-    tipg_api_custom_domain: Optional[str] = Field(
-        description="""Custom domain name for the tipg API.
         Must provide `acm_certificate_arn`""",
         default=None,
     )
@@ -110,12 +109,14 @@ class AppConfig(BaseSettings):
         default=None,
     )
 
-    @pydantic.field_validator("tags")
-    def default_tags(cls, v, info: FieldValidationInfo):
+    model_config = SettingsConfigDict(env_file=".env")
+
+    @field_validator("tags")
+    def default_tags(cls, v, info: ValidationInfo):
         return v or {"project_id": info.data["project_id"], "stage": info.data["stage"]}
 
-    @pydantic.model_validator(mode="after")
-    def validate_nat_gateway_count(self) -> "AppConfig":
+    @model_validator(mode="after")
+    def validate_model(self) -> Self:
         if not self.public_db_subnet and (
             self.nat_gateway_count is not None and self.nat_gateway_count <= 0
         ):
@@ -125,11 +126,7 @@ class AppConfig(BaseSettings):
                              gateways are needed to allow egress from the services
                              and therefore `nat_gateway_count` has to be > 0."""
             )
-        else:
-            return self
 
-    @pydantic.model_validator(mode="after")
-    def validate_stac_browser_version(self) -> "AppConfig":
         if (
             self.stac_browser_version is not None
             and self.stac_api_custom_domain is None
@@ -138,25 +135,21 @@ class AppConfig(BaseSettings):
                 """If a STAC browser version is provided,
                 a custom domain must be provided for the STAC API"""
             )
-        else:
-            return self
 
-    @pydantic.model_validator(mode="after")
-    def validate_acm_certificate_arn(self) -> "AppConfig":
         if self.acm_certificate_arn is None and any(
             [
                 self.stac_api_custom_domain,
-                self.titiler_pgstac_api_custom_domain,
+                self.raster_api_custom_domain,
+                self.vector_api_custom_domain,
                 self.stac_ingestor_api_custom_domain,
-                self.tipg_api_custom_domain,
             ]
         ):
             raise ValueError(
                 """If any custom domain is provided,
                 an ACM certificate ARN must be provided"""
             )
-        else:
-            return self
+
+        return self
 
     def build_service_name(self, service_id: str) -> str:
         return f"{self.project_id}-{self.stage}-{service_id}"
