+++ package.json
@@ -1,8 +1,7 @@
 {
-    "name": "eoapi-template",
-    "version": "0.1.0",
-    "dependencies": {
-      "aws-cdk": "^2.81.0"
-    }
+  "name": "eoapi-template",
+  "version": "0.1.0",
+  "dependencies": {
+    "aws-cdk": "2.130.0"
   }
-  
+}
