+++ requirements.txt
@@ -1,16 +1,6 @@
-aws-cdk-lib>=2.75.0
-aws_cdk.aws_cognito_identitypool_alpha>=2.75.0a0
-aws-cdk.aws-apigatewayv2-alpha==2.95.1a0
-eoapi-cdk==5.4.0
-constructs>=10.0.0,<11.0.0
-pydantic==2.0.2
-pydantic-settings==2.0.1
-black==22.3.0
+eoapi-cdk==7.1.0
+pydantic==2.7
+pydantic-settings==2.2.1
 boto3==1.24.15
-boto3-stubs[cognito-idp,cognito-identity]
-flake8==4.0.1
-click==8.1.3
-requests==2.28.0
-python-dotenv==1.0.0
 pyyaml==6.0
-types-PyYAML==6.0.12.10
+typing-extensions
